move types to own files
