+++ packages/safe-fn/src/util.ts
@@ -1,18 +1,17 @@
 import { ResultAsync, err, ok } from "neverthrow";
 import { z } from "zod";
+import type { AnyRunnableSafeFn } from "./runnable-safe-fn";
 import type {
-  AnyRunnableSafeFn,
-  AnySafeFnCatchHandlerRes,
-  AnySafeFnHandlerRes,
   SafeFnCallBacks,
-  SafeFnInput,
   SafeFnOnCompleteArgs,
   SafeFnOnErrorArgs,
   SafeFnOnSuccessArgs,
-  SafeFnOutput,
-  SafeFnParseError,
-  SafeFnSuperInternalRunReturn,
-} from "./types";
+} from "./types/callbacks";
+import type { AnySafeFnHandlerRes } from "./types/handler";
+import type { SafeFnSuperInternalRunReturn } from "./types/run";
+import type { SafeFnInput, SafeFnOutput } from "./types/schema";
+
+import type { AnySafeFnCatchHandlerRes, SafeFnParseError } from "./types/error";
 
 const NEXT_JS_ERROR_MESSAGES = ["NEXT_REDIRECT", "NEXT_NOT_FOUND"];
